pyproject.toml
[project]
name = "optical-network"
version = "0.1.0"
description = "Optical Network"
dependencies = [
    "numpy>=2.1.0",
    "pandas>=2.2.2",
    "matplotlib>=3.9.2",
    "networkx>=3.3",
    "tqdm>=4.66.5",
    "gurobipy>=12.0.0",
    "ipykernel>=6.29.5",
    "ruff>=0.6.2",
    "hydra-core>=1.3.2",
    "scipy>=1.14.1",
    "openpyxl>=3.1.5",
    "seaborn>=0.13.2",
    "icecream>=2.1.3",
    "jinja2>=3.1.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.11.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/optical_network"]

[tool.ruff]
ignore = ["F403", "F405"]
